Move Command classes to ittybitty.commands package
