+++ src/main/java/ittybotty/commands/ListCommand.java
@@ -1,8 +1,8 @@
-package ittybotty;
+package ittybotty.commands;
 
 public final class ListCommand extends UserCommand {
     @Override
     public String toString() {
-        return "ittybotty.ListCommand{}";
+        return "ittybotty.commands.ListCommand{}";
     }
 }
